code-split : call by fetch_repo_info
